mailer/requirements.txt
Flask==2.2.2
Flask-Mail==0.9.1
Jinja2==3.1.2
pika==1.3.0
requests==2.28.1
python-dotenv==0.20.0
